move transaction fee calculation and initialize fees lazily
