--- crates/astria-sequencer/src/transaction/query.rs
@@ -1,139 +1,0 @@
-use astria_core::{
-    generated::protocol::transaction::v1::TransactionBody as RawBody,
-    protocol::{
-        abci::AbciErrorCode,
-        transaction::v1::TransactionBody,
-    },
-};
-use cnidarium::Storage;
-use prost::{
-    Message as _,
-    Name as _,
-};
-use tendermint::abci::{
-    request,
-    response,
-    Code,
-};
-
-use crate::{
-    app::StateReadExt as _,
-    assets::StateReadExt as _,
-    transaction::checks::get_fees_for_transaction,
-};
-
-pub(crate) async fn transaction_fee_request(
-    storage: Storage,
-    request: request::Query,
-    _params: Vec<(String, String)>,
-) -> response::Query {
-    use astria_core::protocol::fees::v1::TransactionFeeResponse;
-
-    let tx = match preprocess_request(&request) {
-        Ok(tx) => tx,
-        Err(err_rsp) => return err_rsp,
-    };
-
-    // use latest snapshot, as this is a query for a transaction fee
-    let snapshot = storage.latest_snapshot();
-    let height = match snapshot.get_block_height().await {
-        Ok(height) => height,
-        Err(err) => {
-            return response::Query {
-                code: Code::Err(AbciErrorCode::INTERNAL_ERROR.value()),
-                info: AbciErrorCode::INTERNAL_ERROR.info(),
-                log: format!("failed getting block height: {err:#}"),
-                ..response::Query::default()
-            };
-        }
-    };
-
-    let fees_with_ibc_denoms = match get_fees_for_transaction(&tx, &snapshot).await {
-        Ok(fees) => fees,
-        Err(err) => {
-            return response::Query {
-                code: Code::Err(AbciErrorCode::INTERNAL_ERROR.value()),
-                info: AbciErrorCode::INTERNAL_ERROR.info(),
-                log: format!("failed calculating fees for provided transaction: {err:#}"),
-                ..response::Query::default()
-            };
-        }
-    };
-
-    let mut fees = Vec::with_capacity(fees_with_ibc_denoms.len());
-    for (ibc_denom, value) in fees_with_ibc_denoms {
-        let trace_denom = match snapshot.map_ibc_to_trace_prefixed_asset(&ibc_denom).await {
-            Ok(Some(trace_denom)) => trace_denom,
-            Ok(None) => {
-                return response::Query {
-                    code: Code::Err(AbciErrorCode::INTERNAL_ERROR.value()),
-                    info: AbciErrorCode::INTERNAL_ERROR.info(),
-                    log: format!(
-                        "failed mapping ibc denom to trace denom: {ibc_denom}; asset does not \
-                         exist in state"
-                    ),
-                    ..response::Query::default()
-                };
-            }
-            Err(err) => {
-                return response::Query {
-                    code: Code::Err(AbciErrorCode::INTERNAL_ERROR.value()),
-                    info: AbciErrorCode::INTERNAL_ERROR.info(),
-                    log: format!("failed mapping ibc denom to trace denom: {err:#}"),
-                    ..response::Query::default()
-                };
-            }
-        };
-        fees.push((trace_denom.into(), value));
-    }
-
-    let resp = TransactionFeeResponse {
-        height,
-        fees,
-    };
-
-    let payload = resp.into_raw().encode_to_vec().into();
-
-    let height = tendermint::block::Height::try_from(height).expect("height must fit into an i64");
-    response::Query {
-        code: 0.into(),
-        key: request.path.into_bytes().into(),
-        value: payload,
-        height,
-        ..response::Query::default()
-    }
-}
-
-fn preprocess_request(request: &request::Query) -> Result<TransactionBody, response::Query> {
-    let tx = match RawBody::decode(&*request.data) {
-        Ok(tx) => tx,
-        Err(err) => {
-            return Err(response::Query {
-                code: Code::Err(AbciErrorCode::BAD_REQUEST.value()),
-                info: AbciErrorCode::BAD_REQUEST.info(),
-                log: format!(
-                    "failed to decode request data to a protobuf {}: {err:#}",
-                    RawBody::full_name()
-                ),
-                ..response::Query::default()
-            });
-        }
-    };
-
-    let tx = match TransactionBody::try_from_raw(tx) {
-        Ok(tx) => tx,
-        Err(err) => {
-            return Err(response::Query {
-                code: Code::Err(AbciErrorCode::BAD_REQUEST.value()),
-                info: AbciErrorCode::BAD_REQUEST.info(),
-                log: format!(
-                    "failed to convert raw proto unsigned transaction to native unsigned \
-                     transaction: {err:#}"
-                ),
-                ..response::Query::default()
-            });
-        }
-    };
-
-    Ok(tx)
-}
